% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drag.R
\name{tk_drag}
\alias{tk_drag}
\title{Create a tangle kit slider}
\usage{
tk_drag(value, text, min = NULL, max = NULL, id = as_attr(text))
}
\arguments{
\item{value}{Starting value.}

\item{text}{Text shown after the number}

\item{min,max}{Minimum and maximum allowable values.}

\item{id}{A unique identifier for the control. If unspecified, uses
lower case value with non-alphanumeric values converted to -,
multiple - collapsed to one -, and leading and trailing - dropped.}
}
\description{
Create a tangle kit slider
}
\examples{
tk_drag(1, " pies")
tk_drag(2, " hats")
}

